requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
PyPDF2>=3.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tiktoken>=0.5.0
pyyaml>=6.0.1
pathlib>=1.0.1
python-dotenv>=1.0.0
tqdm>=4.66.1
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
pylint>=2.17.5
pandas>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0 